+++ zipkin/doc.go
@@ -1,0 +1,2 @@
+// Package zipkin comprises Zipkin functionality for Zipkin compatiblity.
+package zipkin